Create multiple binaries
Create multiple binaries in the Cargo.toml file.
The two created binaries reference to the main virus program and to the
installer, respectively.
The main virus program is not called virus, just in case Windows try to
go around reading your file names.
A simple file with just the main function was created for the installer.